% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_normalized.R
\name{corr.normalized}
\alias{corr.normalized}
\title{correlation-test with data normalized by RSimNorm.}
\usage{
corr.normalized(
  physeq = NULL,
  count_table = NULL,
  tax_level = NULL,
  meta = NULL,
  main_var,
  type = "pearson",
  method = "BH",
  alpha = 0.05,
  eta = 0,
  lib_cut = 0,
  bootstrap_num = 3
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object which consists of a count table, a sample metadata, a taxonomy table.
The row names of the meta data must match the sample names of the count table.}

\item{count_table}{A numeric matrix or data frame. Row represents taxa. Column represents samples.}

\item{meta}{Meta table.}

\item{main_var}{The name of the continuous variable of interest.}

\item{method}{Correction method for multiple testing effect.}

\item{alpha}{Significance level.}

\item{eta}{The classification error control level. We recommend to choose it between 0 and 0.07.}

\item{lib_cut}{A numerical threshold for filtering samples based on library sizes. Samples with
library sizes less than \code{lib_cut} will be excluded in the analysis. Default is 0, i.e. do not
discard any sample.}

\item{bootstrap_num}{Bootstrap times. Default setting is 3 if the number of taxa is large.}

\item{taxa_level}{Taxonomy level to conduct the test.}
}
\value{
a \code{dataframe} with columns:
\itemize{
\item{ \code{taxa}, a vector. Names of taxa.}
\item{ \code{p-value}, a vector. p-value.}
\item{ \code{adjusted p-value}, a vector. adjusted p-value.}
\item{ \code{is.DA}, a vector. T is differential abundant, F is not differential abundant.}
}
}
\description{
This function conduct differential abundant test by correlation-test with data normalized by RSimNorm.
}
