% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gut_cn.R
\docType{data}
\name{gut_cn}
\alias{gut_cn}
\title{Gut microbiome samples from Guangdong province in China}
\format{
\subsection{\code{gut_cn}}{

A data frame with 37,532 rows and 539 columns:
}
}
\source{
\url{https://qiita.ucsd.edu/study/description/11757}
}
\usage{
gut_cn
}
\description{
A subset of data from gut microbiome of 7009 individuals from 14 district within Guangdong in China.
}
\keyword{datasets}
