% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSimNorm.R
\name{RSimNorm}
\alias{RSimNorm}
\title{A scaling normalization method based on rank correlation.}
\usage{
RSimNorm(
  physeq = NULL,
  count_table = NULL,
  eta = 0.01,
  gamma = 0.8,
  lib_cut = 0,
  bootstrap_num = 3
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object which consists of a count table, a sample metadata. The row
names of the meta data must match the sample names of the count table.}

\item{count_table}{A numeric matrix or data frame. Row represents taxa. Column represents samples.}

\item{eta}{The classification error control level. We recommend to choose it between 0 and 0.07.}

\item{gamma}{The correlation threshold. Recommended value for OTU/ASV level analysis is 0.8.}

\item{lib_cut}{A numerical threshold for filtering samples based on library sizes. Samples with
library sizes less than \code{lib_cut} will be excluded in the analysis. Default is 0, i.e. do not
discard any sample.}

\item{bootstrap_num}{Bootstrap times. Default setting is 3 if the number of taxa is large.}
}
\value{
a \code{list} with components:
\itemize{
\item{ \code{P}, a matrix. Normalized data. Row represents taxa. Column represents samples.}
\item{ \code{I0}, a vector. Selected reference set.}
\item{ \code{pi0}, a number. The estimated proportion of reference set.}
}
}
\description{
A method for efficient normalization of RNA-seq data. This method is designed to eliminate the bias from sequencing depth and reveal the true biological variations. It can be directly applied to PCoA, two-sample test and differential abundant test.
}
\details{
We present a new normalization method on OTUs/ASVs which is robust to zero prevalence
and has a theoretical guarantee of the accuracy of normalization.
}
\examples{

 data(gut_cn)
     Datanorm <- RSimNorm(count_table = A)$P

}
